Convert to C89
* Change -Ofast to -O3 (-Ofast made some FP functions inconsistent
  across platforms because it makes some non-standard optimizations)
* Add pedantic+all+extra warnings, fail on warnings
* Fix over word (was synonym to rot)

+++ readfile.c
@@ -8,7 +8,7 @@ char *readfile(char *filename)
 	FILE *f;
 	int res;
 	char *content;
-	long length;
+	size_t length;
 	size_t readsize;
 
 	content = 0;
